+++ controller/konnect/reconciler_upstreamref.go
@@ -147,11 +147,8 @@ func handleKongUpstreamRef[T constraints.SupportedKonnectEntityType, TEnt constr
 		}
 		if k8serrors.IsNotFound(err) {
 			return ctrl.Result{}, ReferencedControlPlaneDoesNotExistError{
-				Reference: types.NamespacedName{
-					Namespace: ent.GetNamespace(),
-					Name:      cpRef.KonnectNamespacedRef.Name,
-				},
-				Err: err,
+				Reference: cpRef,
+				Err:       err,
 			}
 		}
 		return ctrl.Result{}, err